+++ src/main/java/net/telepathicgrunt/bumblezone/blocks/BzBlocks.java
@@ -86,7 +86,8 @@ public class BzBlocks
     );
     
     public static final ForgeFlowingFluid.Properties SUGAR_WATER_FLUID_PROPERTIES =
-            new ForgeFlowingFluid.Properties(SUGAR_WATER_FLUID, SUGAR_WATER_FLUID_FLOWING, FluidAttributes.Water.builder(FLUID_STILL, FLUID_FLOWING).overlay(FLUID_OVERLAY).viscosity(1500))
+            new ForgeFlowingFluid.Properties(SUGAR_WATER_FLUID, SUGAR_WATER_FLUID_FLOWING, 
+        	    FluidAttributes.Water.builder(FLUID_STILL, FLUID_FLOWING).overlay(FLUID_OVERLAY).viscosity(1500))
                     .bucket(BzItems.SUGAR_WATER_BUCKET).canMultiply().block(SUGAR_WATER_BLOCK);
     
 }