+++ src/main/java/net/telepathicgrunt/bumblezone/blocks/SugarWaterEvents.java
@@ -71,19 +71,22 @@ public class SugarWaterEvents
 
 
 	//*borrowed* from the Item class lol
-	protected static RayTraceResult rayTrace(World worldIn, PlayerEntity player, RayTraceContext.FluidMode fluidMode)
+	protected static RayTraceResult rayTrace(World world, PlayerEntity player, RayTraceContext.FluidMode fluidMode)
 	{
-		float f = player.rotationPitch;
-		float f1 = player.rotationYaw;
-		Vec3d vec3d = player.getEyePosition(1.0F);
-		float f2 = MathHelper.cos(-f1 * ((float) Math.PI / 180F) - (float) Math.PI);
-		float f3 = MathHelper.sin(-f1 * ((float) Math.PI / 180F) - (float) Math.PI);
-		float f4 = -MathHelper.cos(-f * ((float) Math.PI / 180F));
-		float f5 = MathHelper.sin(-f * ((float) Math.PI / 180F));
-		float f6 = f3 * f4;
-		float f7 = f2 * f4;
-		double d0 = player.getAttribute(PlayerEntity.REACH_DISTANCE).getValue();;
-		Vec3d vec3d1 = vec3d.add((double) f6 * d0, (double) f5 * d0, (double) f7 * d0);
-		return worldIn.rayTraceBlocks(new RayTraceContext(vec3d, vec3d1, RayTraceContext.BlockMode.OUTLINE, fluidMode, player));
+		float pitch = player.rotationPitch;
+		float yaw = player.rotationYaw;
+		Vec3d eyePos = player.getEyePosition(1.0F);
+		float f2 = MathHelper.cos(-yaw * ((float) Math.PI / 180F) - (float) Math.PI);
+		float f3 = MathHelper.sin(-yaw * ((float) Math.PI / 180F) - (float) Math.PI);
+		float f4 = -MathHelper.cos(-pitch * ((float) Math.PI / 180F));
+		float yTargetNormalized = MathHelper.sin(-pitch * ((float) Math.PI / 180F));
+		float xTargetNormalized = f3 * f4;
+		float zTargetNormalized = f2 * f4;
+		double targetDistance = player.getAttribute(PlayerEntity.REACH_DISTANCE).getValue();
+		Vec3d targetPos = eyePos.add(
+			(double) xTargetNormalized * targetDistance, 
+			(double) yTargetNormalized * targetDistance, 
+			(double) zTargetNormalized * targetDistance);
+		return world.rayTraceBlocks(new RayTraceContext(eyePos, targetPos, RayTraceContext.BlockMode.OUTLINE, fluidMode, player));
 	}
 }